pyproject.toml
[build-system]
requires = [
    "setuptools>=75.0",
    "build>=1.2.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "mneia-backend"
version = "0.0.1"
dependencies = [
    "django>=5.1",
    "djangorestframework >= 3.15.0",
    "django-cors-headers >= 4.3.1",
    "requests >= 2.31.0",

]
description = "A Django app for Mneia backend."
readme = "README.md"
requires-python = ">= 3.10"
authors = [
    {name = "Marios Zindilis", email = "marios@mneia.gr"},
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
Homepage = "https://mneia.gr/"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "flake8 >= 7.0.0",
    "black >= 24.3.0",
    "isort >= 5.13.2",
    "bandit >= 1.7.8",
    "coverage >= 7.4.4",
    "freezegun >= 1.5.1",
]

[tool.black]
    line-length = 120

[tool.isort]
    multi_line_output = 3  # https://pycqa.github.io/isort/docs/configuration/options.html#multi-line-output
    include_trailing_comma = true  # require trailing comma at the end of a list of imports in parentheses
    line_length = 120  # https://pycqa.github.io/isort/docs/configuration/options.html#line-length

[tool.bandit]
    assert_used = { skips = [ "*/test_*.py" ] } # Allow `assert` to be used in tests

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mneia_backend.tests.settings"
